+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - (Bugfix) Fix ErrorArray String function
 - (Feature) Switch services to Port names
 - (Feature) Configurable ArangoD Port
+- (Bugfix) Do not stop Sync if Synchronization is in progress 
 
 ## [1.2.20](https://github.com/arangodb/kube-arangodb/tree/1.2.20) (2022-10-25)
 - (Feature) Add action progress